requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
requests>=2.26.0
python-dotenv>=1.0.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.7.0
unstructured>=0.10.0
python-multipart>=0.0.6
python-pptx>=0.6.21
python-docx>=0.8.11
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Vector Database and Embeddings
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
torch>=1.9.0
transformers>=4.12.0
chromadb>=0.4.0
langchain>=0.0.300

# Core Dependencies
openai>=1.0.0
langchain-openai>=0.0.5

# Evaluation Metrics
ragas>=0.1.0
rouge>=1.0.1
nltk>=3.8.1

# Data Processing
datasets>=2.14.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.1
loguru>=0.7.2

# Development
tqdm>=4.62.0  # For progress bars